From a36ce6e442b170413d4fb18c5476bc2d2244c004 Mon Sep 17 00:00:00 2001 From: Vincent Rabaud <vrabaud@google.com> Date: Thu, 07 Sep 2023 21:16:03 +0200 Subject: [PATCH] [M108-LTS] Fix OOB write in BuildHuffmanTable.
M108 merge issues:
  dec/vp8l_dec.c:
    - Conflicting checks before ReadHuffmanCodeLengths() return statement;
    In 114, an assignment follows the check instead of a return.
    - ReadHuffmanCodes(): Conflict after the changed huffman_tables check,
    there's an assignment in 114 instead of the setter call.

First, BuildHuffmanTable is called to check if the data is valid.
If it is and the table is not big enough, more memory is allocated.

This will make sure that valid (but unoptimized because of unbalanced
codes) streams are still decodable.

Bug: chromium:1479274
Change-Id: I31c36dbf3aa78d35ecf38706b50464fd3d375741
(cherry picked from commit 902bc9190331343b2017211debcec8d2ab87e17a)

Signed-off-by: dubj <dubingjian@huawei.com>
Change-Id: If1a2bf7435309a6e97dacdef6a67fd30e4d74579